[REF-5957] implement SuppressExceptions decorator
The idea is that if we have an API that throws exception,
we could easily control it: suppress exception or do some other action.

Consider we have: set_stat().
No we can add a decorator:

@SuppressException(MLOpsException)
set_stat()

Now, if set_stat() throws exception, it will be processed in decorator
function, according to the rules.